requirements.txt
pycocotools==2.0.1
torch==1.5.1
torchvision==0.6.1
PyYAML==5.3.1
streamlit==0.62.1
